requirements.txt
Flask==3.0.3
openai==1.37.0
python-dotenv==0.19.2
requests==2.26.0
